requirements.txt
gradio
ollama
PyPDF2
Pillow
python-dotenv
